% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\docType{data}
\name{mapping.old}
\alias{mapping.old}
\title{run CATE analysis}
\format{
An object of class \code{data.frame} with 138291 rows and 5 columns.
}
\usage{
mapping.old
}
\arguments{
\item{input.path}{string with the path to the folder containing phase 2.2 data}

\item{output.path}{string with the path to the folder to save the results}

\item{siteid}{string}
}
\value{
dataframe containing the results for all outcomes of interest
}
\description{
run CATE analysis
}
\keyword{datasets}
