% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{run CATE analysis}
\usage{
run_analysis(input.path, output.path, siteid = "VA")
}
\arguments{
\item{input.path}{string with the path to the folder containing phase 2.2 data}

\item{output.path}{string with the path to the folder to save the results}

\item{siteid}{string}
}
\value{
dataframe containing the results for all outcomes of interest
}
\description{
run CATE analysis
}
