% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_estimation.R
\name{site_estimation}
\alias{site_estimation}
\title{Estimate CATE using only data in each site}
\usage{
site_estimation(X, A, Y, threshold = 0)
}
\arguments{
\item{X}{matrix containing all binary covariates}

\item{A}{matrix containing binary vaccine information}

\item{Y}{matrix containing binary outcome}

\item{threshold}{integer to only count covariates with enough occurences}
}
\description{
Estimate CATE using only data in each site
}
